refactored frontend-files
